Test that isNull() and isNotNull() work even on types whose equals() implementations throw NPE when given null.
I don't really "want" to support this, but tests in the depot currently rely on isNull() and isNotNull() to work with broken equals() implementations, and my upcoming changes would causes their tests to fail. It would be nice to clean them up, but accommodating them (starting with this commit to verify that we've done so) is a more expedient way forward.

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=187175835